% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useNThreads.R
\name{allocateJobs}
\alias{allocateJobs}
\title{allocateJobs}
\usage{
allocateJobs(nTasks, nWorkers)
}
\arguments{
\item{nWorkers}{}
}
\value{

}
\description{
Facilitates multi-threading by producing an even allocation of jobs
  Works even when number of jobs is less than number of threads in which case some components of the
  returned allocation will have length 0.
}
