% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiData.R
\name{as.list.multiSet}
\alias{as.list.multiSet}
\title{Convert a multiSet object to a list}
\usage{
as.list.multiSet(x, ...)
}
\arguments{
\item{x}{muiltiSet object to be converted to list}

\item{\dots}{other arguments passed to multiSet2list}
}
\description{
Each dataset is a list with a data element with the expression.
}
