% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiData.R
\name{as.multiSet.list}
\alias{as.multiSet.list}
\title{Convert a list to a multiSet}
\usage{
as.multiSet.list(multiSet)
}
\arguments{
\item{multiSet}{list to be converted to a multiSet object, or a multiSet to
be converted to a list}

\item{\dots}{other arguments to be passed to \code{list2multiSet} or
\code{list2multiset}}
}
\value{
An object of class multiSet or a list.
}
\description{
Convert a list to a multiSet
}
\seealso{
\code{\link{list2multiSet}}, \code{\link{multiSet2list}}
}
