% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{automaticNetworkScreeningGS}
\alias{automaticNetworkScreeningGS}
\title{One-step automatic network gene screening with external gene significance}
\usage{
automaticNetworkScreeningGS(datExpr, GS, power = 6,
  networkType = "unsigned", detectCutHeight = 0.995,
  minModuleSize = min(20, ncol(as.matrix(datExpr))/2), datME = NULL)
}
\arguments{
\item{datExpr}{data frame containing the expression data, columns
corresponding to genes and rows to samples}

\item{GS}{vector containing gene significance for all genes given in
\code{datExpr}}

\item{power}{soft thresholding power used in network construction}

\item{networkType}{character string specifying network type. Allowed values
are (unique abbreviations of) \code{"unsigned"}, \code{"signed"},
\code{"hybrid"}.}

\item{detectCutHeight}{cut height of the gene hierarchical clustering
dendrogram. See \code{cutreeDynamic} for details.}

\item{minModuleSize}{minimum module size to be used in module detection
procedure.}

\item{datME}{optional specification of module eigengenes. A data frame whose
columns are the module eigengenes. If given, module analysis will not be
performed.}

\item{datKME}{a data frame (or a matrix-like object) containing
eigengene-based connectivities of all genes in the network.}

\item{GS}{a vector with one entry for every gene containing its gene
significance.}
}
\value{
A list with the following components: \item{networkScreening}{a data
frame containing results of the network screening procedure. See
\code{\link{networkScreeningGS}} for more details.} \item{datME}{ calculated
module eigengenes (or a copy of the input \code{datME}, if given).}
\item{hubGeneSignificance}{ hub gene significance for all calculated
modules. See \code{\link{hubGeneSignificance}}. }

A vector whose entries are the hub gene significances for each
module.
}
\description{
This function performs gene screening based on external gene significance
and their network properties.
}
\details{
Network screening is a method for identifying genes that have a high gene
significance and are members of important modules at the same time.  If
\code{datME} is given, the function calls \code{\link{networkScreeningGS}}
with the default parameters. If \code{datME} is not given, module eigengenes
are first calculated using network analysis based on supplied parameters.
}
\author{
Steve Horvath

Steve Horvath
}
\references{
Dong J, Horvath S (2007) Understanding Network Concepts in
Modules, BMC Systems Biology 2007, 1:24
}
\seealso{
\code{\link{networkScreening}}, \code{\link{hubGeneSignificance}},
\code{\link{networkScreening}}, \code{\link[dynamicTreeCut]{cutreeDynamic}}
}
\keyword{Calculate}
\keyword{Hubgene}
\keyword{In}
\keyword{\code{datKME}}
\keyword{all}
\keyword{and}
\keyword{approximate}
\keyword{columns}
\keyword{correspond}
\keyword{for}
\keyword{gene}
\keyword{genes}
\keyword{hub}
\keyword{in}
\keyword{misc}
\keyword{modules}
\keyword{modules.}
\keyword{network.}
\keyword{rows}
\keyword{significance}
\keyword{to}

