% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useNThreads.R
\name{blockSize}
\alias{blockSize}
\title{blockSize}
\usage{
blockSize(
  matrixSize,
  rectangularBlocks = TRUE,
  maxMemoryAllocation = NULL,
  overheadFactor = 3
)
}
\arguments{
\item{overheadFactor}{}
}
\value{

}
\description{
Function to calculate an appropriate blocksize
}
