% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockwiseModulesC.R
\name{blockwiseModules}
\alias{blockwiseModules}
\title{blockwiseModules}
\usage{
blockwiseModules(
  datExpr,
  weights = NULL,
  checkMissingData = TRUE,
  blocks = NULL,
  maxBlockSize = 5000,
  blockSizePenaltyPower = 5,
  nPreclusteringCenters = as.integer(min(ncol(datExpr)/20, 100 *
    ncol(datExpr)/maxBlockSize)),
  randomSeed = 12345,
  loadTOM = FALSE,
  corType = "pearson",
  maxPOutliers = 1,
  quickCor = 0,
  pearsonFallback = "individual",
  cosineCorrelation = FALSE,
  power = 6,
  networkType = "unsigned",
  replaceMissingAdjacencies = FALSE,
  suppressTOMForZeroAdjacencies = FALSE,
  TOMType = "signed",
  TOMDenom = "min",
  getTOMs = NULL,
  saveTOMs = FALSE,
  saveTOMFileBase = "blockwiseTOM",
  deepSplit = 2,
  detectCutHeight = 0.995,
  minModuleSize = min(20, ncol(datExpr)/2),
  maxCoreScatter = NULL,
  minGap = NULL,
  maxAbsCoreScatter = NULL,
  minAbsGap = NULL,
  minSplitHeight = NULL,
  minAbsSplitHeight = NULL,
  useBranchEigennodeDissim = FALSE,
  minBranchEigennodeDissim = mergeCutHeight,
  stabilityLabels = NULL,
  stabilityCriterion = c("Individual fraction", "Common fraction"),
  minStabilityDissim = NULL,
  pamStage = TRUE,
  pamRespectsDendro = TRUE,
  reassignThreshold = 1e-06,
  minCoreKME = 0.5,
  minCoreKMESize = minModuleSize/3,
  minKMEtoStay = 0.3,
  mergeCutHeight = 0.15,
  impute = TRUE,
  trapErrors = FALSE,
  numericLabels = FALSE,
  nThreads = 0,
  useInternalMatrixAlgebra = FALSE,
  useCorOptionsThroughout = TRUE,
  verbose = 0,
  indent = 0,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Function to calculate modules and eigengenes from all genes.
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[fastcluster]{hclust}}
}
