\name{clusterCoef}
\alias{clusterCoef}
\title{ Clustering coefficient calculation }
\description{
  This function calculates the clustering coefficients for all nodes in the network given by the input
adjacency matrix. 
}
\usage{
clusterCoef(adjMat)
}
\arguments{
  \item{adjMat}{ adjacency matrix }
}
\value{
  A vector of clustering coefficients for each node.
}
\author{ Steve Horvath }
\keyword{ misc }
