\name{collectGarbage}
\alias{collectGarbage}
\title{Iterative garbage collection. }
\description{
  Performs garbage collection until free memory idicators show no change.
}
\usage{
collectGarbage()
}
\value{
  None.
}
\author{ Steve Horvath }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{utilities}
