% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiData.R
\name{colnames<-.multiSet}
\alias{colnames<-.multiSet}
\title{Get and set column names in a multiSet structure.}
\usage{
\method{colnames}{multiSet}(multiSet) <- value
}
\arguments{
\item{multiSet}{A multiSet structure}

\item{value}{A vector (coercible to character) of column names.}
}
\value{
The multiSet structure with the column names
set in all \code{data} components.
}
\description{
Get and set column names on each \code{data} component in a multiSet
structure.
}
\details{
A multiSet structure is intended to store (the same type of) data for
multiple, possibly independent, realizations (for example, expression data
for several independent experiments). It is a list where each component
corresponds to an (independent) data set. Each component is in turn a list
that can hold various types of information but must have a \code{data}
component. In a "strict" multiSet structure, the \code{data} components are
required to each be a matrix or a data frame and have the same number of
columns. In a "loose" multiSet structure, the \code{data} components can be
anything (but for most purposes should be of comparable type and content).

The \code{colnames.multiSet}  assume (and checks for) a"strict" multiSet
structure.
}
\author{
Peter Langfelder
}
\seealso{
\code{\link{multiSet}} to create a multiSet structure.
}
\keyword{misc}

