% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiData.R
\name{colnames<-}
\alias{colnames}
\alias{colnames.default}
\alias{colnames.multiSet}
\alias{colnames<-}
\alias{colnames<-.default}
\title{Extract the colnames of the multiSet}
\usage{
colnames(x) <- value

\method{colnames}{default}(x, ...) <- value

colnames(x)

\method{colnames}{default}(obj, ...)

\method{colnames}{multiSet}(multiSet)
}
\arguments{
\item{obj}{object to extract the colnames}

\item{multiSet}{multiSet object to retrieve the columns of the first data}

\item{\dots}{other arguments passed}
}
\value{
\code{colnames.multiSet} returns the vector of column names of the
\code{data} component. The function assumes the column names in all sets are
the same.
}
\description{
Extract the colnames of the multiSet
}

