% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulePreservation.R
\name{.accuracyStatistics}
\alias{.accuracyStatistics}
\title{.accuracyStatistics}
\usage{
.accuracyStatistics(colorRef, colorTest, ccTupletSize, greyName, pEpsilon)
}
\arguments{
\item{colorRef}{}

\item{colorTest}{}

\item{ccTupletSize}{}

\item{greyName}{}

\item{pEpsilon}{}
}
\value{

}
\description{
.accuracyStatistics
}
