% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulePreservation.R
\name{.checkAdj}
\alias{.checkAdj}
\title{.checkAdj}
\usage{
.checkAdj(multiAdj, verbose, indent)
}
\arguments{
\item{multiAdj}{}

\item{verbose}{}

\item{indent}{}
}
\value{

}
\description{
.checkAdj
}
