% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulePreservation.R
\name{.checkExpr}
\alias{.checkExpr}
\title{.checkExpr}
\usage{
.checkExpr(multiExpr, verbose, indent)
}
\arguments{
\item{multiExpr}{}

\item{verbose}{}

\item{indent}{}
}
\value{

}
\description{
.checkExpr
}
