% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulePreservation.R
\name{.clusterCoeff}
\alias{.clusterCoeff}
\title{.clusterCoeff}
\usage{
.clusterCoeff(adjacency)

.clusterCoeff(adjacency)
}
\arguments{
\item{adjacency}{}

\item{adjmat1}{}
}
\value{



}
\description{
.clusterCoeff

.clusterCoeff
}
