% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulePreservation.R
\name{.computeLinksInNeighbors}
\alias{.computeLinksInNeighbors}
\title{.computeLinksInNeighbors}
\usage{
.computeLinksInNeighbors(x, imatrix)
}
\arguments{
\item{x}{}

\item{imatrix}{}
}
\value{

}
\description{
.computeLinksInNeighbors
}
