% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulePreservation.R
\name{.coreCalcForAdj}
\alias{.coreCalcForAdj}
\title{.coreCalcForAdj}
\usage{
.coreCalcForAdj(datRef, datRefP, datTest, colors, opt)
}
\arguments{
\item{datRef}{}

\item{datRefP}{}

\item{datTest}{}

\item{colors}{}

\item{opt}{}
}
\value{

}
\description{
.coreCalcForAdj
}
