% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulePreservation.R
\name{.coreCalcForExpr}
\alias{.coreCalcForExpr}
\title{.coreCalcForExpr}
\usage{
.coreCalcForExpr(datRef, datRefP, datTest, colors, opt)
}
\arguments{
\item{datRef}{}

\item{datRefP}{}

\item{datTest}{}

\item{colors}{}

\item{opt}{}
}
\value{

}
\description{
.coreCalcForExpr
}
