% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulePreservation.R
\name{.modulePreservationInternal}
\alias{.modulePreservationInternal}
\title{.modulePreservationInternal}
\usage{
.modulePreservationInternal(
  multiData,
  multiColor,
  dataIsExpr,
  calculatePermutation,
  multiColorForAccuracy = NULL,
  networkType = "signed",
  corFnc = "cor",
  corOptions = "use = 'p'",
  referenceNetworks = 1,
  testNetworks,
  densityOnly = FALSE,
  maxGoldModuleSize = 1000,
  maxModuleSize = 1000,
  quickCor = 1,
  ccTupletSize,
  calculateCor.kIMall,
  calculateClusterCoeff,
  checkData = TRUE,
  greyName,
  pEpsilon = 1e-200,
  verbose = 1,
  indent = 0
)
}
\arguments{
\item{multiData}{}

\item{multiColor}{}

\item{dataIsExpr}{}

\item{calculatePermutation}{}

\item{multiColorForAccuracy}{}

\item{networkType}{}

\item{corFnc}{}

\item{corOptions}{}

\item{referenceNetworks}{}

\item{testNetworks}{}

\item{densityOnly}{}

\item{maxGoldModuleSize}{}

\item{maxModuleSize}{}

\item{quickCor}{}

\item{ccTupletSize}{}

\item{calculateCor.kIMall}{}

\item{calculateClusterCoeff}{}

\item{checkData}{}

\item{greyName}{}

\item{pEpsilon}{}

\item{verbose}{}

\item{indent}{}
}
\value{

}
\description{
.modulePreservationInternal
}
