% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulePreservation.R
\name{.pValueFromZ}
\alias{.pValueFromZ}
\title{.pValueFromZ}
\usage{
.pValueFromZ(Z, bonf = FALSE, summaryCols = NULL, summaryInd = NULL)
}
\arguments{
\item{Z}{}

\item{bonf}{}

\item{summaryCols}{}

\item{summaryInd}{}
}
\value{

}
\description{
.pValueFromZ
}
