% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulePreservation.R
\name{.qValueFromP}
\alias{.qValueFromP}
\title{.qValueFromP}
\usage{
.qValueFromP(p, summaryCols = NULL, summaryInd = NULL)
}
\arguments{
\item{p}{}

\item{summaryCols}{}

\item{summaryInd}{}
}
\value{

}
\description{
.qValueFromP
}
