\name{GOenrichmentAnalysis}
\alias{GOenrichmentAnalysis}
\title{ Calculation of GO enrichment}
\description{
NOTE: GOenrichmentAnalysis is now a defunct stub that just prints an informational message. Please use function
          enrichmentAnalysis.Entrez from R package anRichment, available from
          https://github.com/plangfelder/anRichment.
}
\usage{
GOenrichmentAnalysis(...)
}
\arguments{
  \item{...}{ Ignored. }
}
\author{ Peter Langfelder }
\seealso{ Replacement functions are available from R package anRichment at https://github.com/plangfelder/anRichment }

\keyword{ misc }% __ONLY ONE__ keyword per line
