% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiData.R
\name{list2multiSet}
\alias{as.multiSet}
\alias{as.multiSet.default}
\alias{as.multiSet.default.list}
\alias{list2multiSet}
\alias{multiSet2list}
\title{Convert a list to a multiSet structure and vice-versa.}
\usage{
list2multiSet(data)

as.multiSet(data)

\method{as.multiSet.default}{list}(x)

as.multiSet.default(x, ...)

multiSet2list(multiSet)
}
\arguments{
\item{data}{A list to be converted to a multiSet structure.}

\item{x}{Object to convert to multiSet or to list}

\item{multiSet}{A multiSet structure to be converted to a list.}

\item{\dots}{Other arguments}
}
\value{
For \code{list2multiSet}, a multiSet structure; for
\code{multiSet2list}, the corresponding list.
}
\description{
\code{list2multiSet} converts a list to a multiSet structure
\code{multiSet2list} does the inverse.
}
\details{
A multiSet structure is a vector of lists (one list for each set) where
each list has a component \code{data} containing some useful information.
}
\author{
Peter Langfelder
}
\keyword{misc}

