% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{moduleEigengenes}
\alias{moduleEigengenes}
\title{moduleEigengenes}
\usage{
moduleEigengenes(
  expr,
  colors,
  impute = TRUE,
  nPC = 1,
  align = "along average",
  excludeGrey = FALSE,
  grey = if (is.numeric(colors)) 0 else "grey",
  subHubs = TRUE,
  trapErrors = FALSE,
  returnValidOnly = trapErrors,
  softPower = 6,
  scale = TRUE,
  verbose = 0,
  indent = 0
)
}
\arguments{
\item{expr}{}

\item{colors}{}

\item{impute}{}

\item{nPC}{}

\item{align}{}

\item{excludeGrey}{}

\item{grey}{}

\item{subHubs}{}

\item{trapErrors}{}

\item{returnValidOnly}{}

\item{softPower}{}

\item{scale}{}

\item{verbose}{}

\item{indent}{}
}
\value{

}
\description{
finds the first principal component (eigengene) in each
module defined by the colors of the input vector "colors".
The theoretical underpinnings are described in Horvath, Dong, Yip (2005)
http://www.genetics.ucla.edu/labs/horvath/ModuleConformity/
This requires the R library impute
}
