% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulePreservation.R
\name{modulePreservation}
\alias{modulePreservation}
\title{modulePreservation}
\usage{
modulePreservation(
  multiData,
  multiColor,
  dataIsExpr = TRUE,
  networkType = "unsigned",
  corFnc = "cor",
  corOptions = "use = 'p'",
  referenceNetworks = 1,
  testNetworks = NULL,
  nPermutations = 100,
  includekMEallInSummary = FALSE,
  restrictSummaryForGeneralNetworks = TRUE,
  calculateQvalue = FALSE,
  randomSeed = 12345,
  maxGoldModuleSize = 1000,
  maxModuleSize = 1000,
  quickCor = 1,
  ccTupletSize = 2,
  calculateCor.kIMall = FALSE,
  calculateClusterCoeff = FALSE,
  useInterpolation = FALSE,
  checkData = TRUE,
  greyName = NULL,
  savePermutedStatistics = TRUE,
  loadPermutedStatistics = FALSE,
  permutedStatisticsFile = if (useInterpolation) "permutedStats-intrModules.RData" else
    "permutedStats-actualModules.RData",
  plotInterpolation = TRUE,
  interpolationPlotFile = "modulePreservationInterpolationPlots.pdf",
  discardInvalidOutput = TRUE,
  parallelCalculation = FALSE,
  verbose = 1,
  indent = 0
)
}
\arguments{
\item{multiData}{}

\item{multiColor}{}

\item{dataIsExpr}{}

\item{networkType}{}

\item{corFnc}{}

\item{corOptions}{}

\item{referenceNetworks}{}

\item{testNetworks}{}

\item{nPermutations}{}

\item{includekMEallInSummary}{}

\item{restrictSummaryForGeneralNetworks}{}

\item{calculateQvalue}{}

\item{randomSeed}{}

\item{maxGoldModuleSize}{}

\item{maxModuleSize}{}

\item{quickCor}{}

\item{ccTupletSize}{}

\item{calculateCor.kIMall}{}

\item{calculateClusterCoeff}{}

\item{useInterpolation}{}

\item{checkData}{}

\item{greyName}{}

\item{savePermutedStatistics}{}

\item{loadPermutedStatistics}{}

\item{permutedStatisticsFile}{}

\item{plotInterpolation}{}

\item{interpolationPlotFile}{}

\item{discardInvalidOutput}{}

\item{parallelCalculation}{}

\item{verbose}{}

\item{indent}{}
}
\value{

}
\description{
modulePreservation
}
