% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiData.R
\name{multiSet.rbindSelf}
\alias{multiSet.rbindSelf}
\title{Turn a multiSet structure into a single matrix or data frame.}
\usage{
multiSet.rbindSelf(multiSet)
}
\arguments{
\item{multiSet}{A multiSet structure.}
}
\value{
A single matrix or data frame containing the "rbinded" result.
}
\description{
This function "rbinds" the \code{data} components of all sets in the input
into a single matrix or data frame.
}
\details{
A multiSet structure is intended to store (the same type of) data for
multiple, possibly independent, realizations (for example, expression data
for several independent experiments). It is a list where each component
corresponds to an (independent) data set. Each component is in turn a list
that can hold various types of information but must have a \code{data}
component. In a "strict" multiSet structure, the \code{data} components are
required to each be a matrix or a data frame and have the same number of
columns. In a "loose" multiSet structure, the \code{data} components can be
anything (but for most purposes should be of comparable type and content).

This function requires a "strict" multiSet structure.
}
\seealso{
\code{\link{multiSet}} to create a multiSet structure

\code{\link{rbind}} for various subtleties of the row binding operation.
}
\author{
Peter Langfelder
}
\keyword{misc}
