% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiData.R
\name{multiSet}
\alias{multiSet}
\title{Create a multiSet structure.}
\usage{
multiSet(...)
}
\arguments{
\item{\dots}{Arguments to be stored in the multiSet structure.}
}
\value{
The resulting multiSet structure.
}
\description{
This function creates a multiSet structure by storing its input arguments
as the 'data' components.
}
\details{
A multiSet structure is intended to store (the same type of) data for
multiple, possibly independent, realizations (for example, expression data
for several independent experiments). It is a list where each component
corresponds to an (independent) data set. Each component is in turn a list
that can hold various types of information but must have a \code{data}
component. In a "strict" multiSet structure, the \code{data} components are
required to each be a matrix or a data frame and have the same number of
columns. In a "loose" multiSet structure, the \code{data} components can be
anything (but for most purposes should be of comparable type and content).
}
\examples{

data1 <- matrix(rnorm(100), 20, 5)
data2 <- matrix(rnorm(50), 10, 5)

md <- multiSet(Set1 = data1, Set2 = data2)

checkSets(md)
}
\author{
Peter Langfelder
}
\seealso{
\code{\link{multiSet2list}} for converting a multiSet structure
to a list; \code{\link{list2multiSet}} for an alternative way of creating a
multiSet structure; \code{\link{apply.multiSet}, \link{apply.multiSetToSubset},
\link{multiSet.mapply}} for ways of applying a function to each component of a
multiSet structure.
}
\keyword{misc}

