% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiData.R
\name{multiSet.setAttr}
\alias{multiSet.setAttr}
\title{Set attributes on each component of a multiSet structure}
\usage{
multiSet.setAttr(multiSet, attribute, valueList)
}
\arguments{
\item{multiSet}{A multiSet structure.}

\item{attribute}{Name for the attribute to be set}

\item{valueList}{List that gives the attribute value for each set in the
multiSet structure.}
}
\value{
The input multiSet with the attribute set on each \code{data}
component.
}
\description{
Set attributes on each \code{data} component of a multiSet structure
}
\seealso{
\code{\link{multiSet}} to create a multiSet structure

\code{is.multiSet} for a description of the multiSet structure.
}
\author{
Peter Langfelder
}
\keyword{misc}
