\name{multiUnion}
\alias{multiUnion}
\alias{multiIntersect}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Union and intersection of multiple sets
}
\description{
Union and intersection of multiple sets. These function generalize the standard functions
\code{\link{union}} and \code{\link{intersect}}.
}
\usage{
multiUnion(setList)
multiIntersect(setList)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{setList}{
A list containing the sets to be performed upon.
}
}
\value{
  The union or intersection of the given sets.
}
\author{
Peter Langfelder
}

\seealso{
The "standard" functions \code{\link{union}} and \code{\link{intersect}}.
}
\keyword{misc}% __ONLY ONE__ keyword per line
