\name{nPresent}
\alias{nPresent}
\title{ Number of present data entries. }
\description{
  A simple sum of present entries in the argument.
}
\usage{
nPresent(x)
}
\arguments{
  \item{x}{ data in which to count number of present entries. }
}
\value{
  A single number giving the number of present entries in \code{x}.
}
\author{ Steve Horvath }
\keyword{ misc }% __ONLY ONE__ keyword per line
