% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{pickSoftThreshold}
\alias{pickSoftThreshold}
\title{pickSoftThreshold}
\usage{
pickSoftThreshold(
  data,
  dataIsExpr = TRUE,
  weights = NULL,
  RsquaredCut = 0.85,
  powerVector = c(seq(1, 10, by = 1), seq(12, 20, by = 2)),
  removeFirst = FALSE,
  nBreaks = 10,
  blockSize = NULL,
  corFnc = cor,
  corOptions = list(use = "p"),
  networkType = "unsigned",
  moreNetworkConcepts = FALSE,
  gcInterval = NULL,
  verbose = 0,
  indent = 0
)
}
\arguments{
\item{data}{}

\item{dataIsExpr}{}

\item{weights}{}

\item{RsquaredCut}{}

\item{powerVector}{}

\item{removeFirst}{}

\item{nBreaks}{}

\item{blockSize}{}

\item{corFnc}{}

\item{corOptions}{}

\item{networkType}{}

\item{moreNetworkConcepts}{}

\item{gcInterval}{}

\item{verbose}{}

\item{indent}{}
}
\value{

}
\description{
The function pickSoftThreshold allows one to estimate the power
 parameter when using a soft thresholding approach with the use of the power
 function \eqn{AF(s)=s^Power}. The removeFirst option removes the first point
 (k=1, P(k=1)) from the regression fit.
}
