\name{qvalue.restricted}
\alias{qvalue.restricted}
\title{
qvalue convenience wrapper
}
\description{
This function calls \code{\link{qvalue}} on finite input p-values, optionally traps errors from the q-value
calculation, and returns just the q values.
}
\usage{
qvalue.restricted(p, trapErrors = TRUE, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{p}{
 a vector of p-values. Missing data are allowed and will be removed.
}
  \item{trapErrors}{
 logical: should errors generated by function \code{\link{qvalue}} trapped? If \code{TRUE}, the errors will be
silently ignored and the returned q-values will all be \code{NA}.
}
  \item{\dots}{
  other arguments to function \code{\link{qvalue}}.
}
}
\value{
  A vector of q-values. Entries whose corresponding p-values were not finite will be \code{NA}.  
}
\author{
Peter Langfelder
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{qvalue}}
}
\keyword{misc}
