% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiData.R
\name{simplify.multiSet}
\alias{simplify.multiSet}
\title{If possible, simplify a multiSet structure to a 3-dimensional array.}
\usage{
\method{simplify}{multiSet}(multiSet)
}
\arguments{
\item{multiSet}{A multiSet structure in the "strict" sense (see below).}
}
\value{
A 3-dimensional array collecting all \code{data} components.
}
\description{
This function attempts to put all \code{data} components into a
3-dimensional array, with the last dimension corresponding to the sets. This
is only possible if all \code{data} components are matrices or data frames
with the same dimensiosn.
}
\details{
A multiSet structure is intended to store (the same type of) data for
multiple, possibly independent, realizations (for example, expression data
for several independent experiments). It is a list where each component
corresponds to an (independent) data set. Each component is in turn a list
that can hold various types of information but must have a \code{data}
component. In a "strict" multiSet structure, the \code{data} components are
required to each be a matrix or a data frame and have the same number of
columns. In a "loose" multiSet structure, the \code{data} components can be
anything (but for most purposes should be of comparable type and content).

This function assumes a "strict" multiSet structure.
}
\note{
The function is relatively fragile and may fail. Use at your own risk.
}
\author{
Peter Langfelder
}
\seealso{
\code{\link{multiSet}} to create a multiSet structure

\code{\link{multiSet2list}} for converting multiSet structures to plain
lists.
}
\keyword{misc}

