\name{sizeGrWindow}
\alias{sizeGrWindow}
\title{ Opens a graphics window with specified dimensions }
\description{
  If a graphic device window is already open, it is closed and re-opened with specified dimensions (in
inches); otherwise a new window is opened.
}
\usage{
sizeGrWindow(width, height)
}
\arguments{
  \item{width}{ desired width of the window, in inches. }
  \item{height}{ desired heigh of the window, in inches. }
}
\value{
  None.
}
\author{ Peter Langfelder }
\keyword{ misc }
