\name{stdErr}
\alias{stdErr}
\title{ Standard error of the mean of a given vector. }
\usage{
stdErr(x)
}
\description{
  Returns the standard error of the mean of a given vector. Missing values are ignored.
}
\arguments{
  \item{x}{ a numeric vector }
}
\value{
  Standard error of the mean of x.
}
\author{ Steve Horvath }
\keyword{ misc }% __ONLY ONE__ keyword per line
