% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiData.R
\name{subset.multiSet}
\alias{subset.multiSet}
\alias{subset}
\alias{subset.default}
\alias{[.multiSet}
\alias{subset}
\alias{subset.default}
\alias{[.multiSet}
\alias{subset.multiSet}
\title{Subset rows and columns in a multiSet structure}
\usage{
\method{subset}{multiSet}(multiSet, rowIndex = NULL, colIndex = NULL,
  permissive = FALSE, drop = FALSE, ...)

subset(multiSet, ...)

\method{subset}{default}(multiSet, ...)

\method{[}{multiSet}(x, i, j, drop = TRUE)
}
\arguments{
\item{multiSet}{A multiSet structure.}

\item{rowIndex}{A list in which each component corresponds to a set and is a
vector giving the rows to be retained in that set. All indexing methods
recognized by R can be used (numeric, logical, negative indexing, etc). If
\code{NULL}, all columns will be retained in each set. Note that setting
individual elements of \code{rowIndex} to \code{NULL} will lead to errors.}

\item{colIndex}{A vector giving the columns to be retained. All indexing
methods recognized by R can be used (numeric, logical, negative indexing,
etc). In addition, column names of the retained columns may be given; if a
given name cannot be matched to a column, an error will be thrown. If
\code{NULL}, all columns will be retained.}

\item{permissive}{logical: should the function tolerate "loose"
\code{multiSet} input? Note that the subsetting may lead to cryptic errors
if the input \code{multiSet} does not follow the "strict" format.}

\item{drop}{logical: should dimensions with extent 1 be dropped?}

\item{\dots}{Other arguments}

\item{x}{MultiSet Object to subset}

\item{i}{Columns to subset of the object}

\item{j}{Rows to subset of the object}

\item{i}{Columns to subset of the object}

\item{j}{Rows to subset of the object}

\item{x}{MultiSet Object to subset}
}
\value{
A multiSet structure containing the selected rows and columns. Note
that result always retaines it dimension and other attributes.
}
\description{
The function restricts each \code{data} component to the given columns and
rows.
}
\details{
A multiSet structure is intended to store (the same type of) data for
multiple, possibly independent, realizations (for example, expression data
for several independent experiments). It is a list where each component
corresponds to an (independent) data set. Each component is in turn a list
that can hold various types of information but must have a \code{data}
component. In a "strict" multiSet structure, the \code{data} components are
required to each be a matrix or a data frame and have the same number of
columns. In a "loose" multiSet structure, the \code{data} components can be
anything (but for most purposes should be of comparable type and content).

This function assumes a "strict" multiSet structure unless \code{permissive}
is \code{TRUE}.
}
\examples{


data1 <- matrix(rnorm(100), 20, 5)
data2 <- matrix(rnorm(50), 10, 5)

md <- multiSet(Set1 = data1, Set2 = data2)
md[, c(1,3)]
md[c(1,3), ]

}
\seealso{
\code{\link{multiSet}} to create a multiSet structure.
}
\author{
Peter Langfelder
}
\keyword{misc}
