% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockwiseModulesC.R
\name{TOMdist}
\alias{TOMdist}
\title{TOMdist}
\usage{
TOMdist(
  adjMat,
  TOMType = unsigned,
  TOMDenom = min,
  suppressTOMForZeroAdjacencies = FALSE,
  useInternalMatrixAlgebra = FALSE,
  verbose = 1,
  indent = 0
)
}
\arguments{
\item{adjMat}{PARAM_DESCRIPTION}

\item{TOMType}{PARAM_DESCRIPTION, Default: unsigned}

\item{TOMDenom}{PARAM_DESCRIPTION, Default: min}

\item{suppressTOMForZeroAdjacencies}{PARAM_DESCRIPTION, Default: FALSE}

\item{useInternalMatrixAlgebra}{PARAM_DESCRIPTION, Default: FALSE}

\item{verbose}{PARAM_DESCRIPTION, Default: 1}

\item{indent}{PARAM_DESCRIPTION, Default: 0}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}

}
