% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{addErrorBars}
\alias{addErrorBars}
\title{Add error bars to a barplot.}
\usage{
addErrorBars(means, errors, two.side = FALSE)
}
\arguments{
\item{means}{vector of means plotted in the barplot}

\item{errors}{vector of standard errors (signle positive values) to be
plotted.}

\item{two.side}{should the error bars be two-sided?}
}
\value{
None.
}
\description{
This function adds error bars to an existing barplot.
}
\author{
Steve Horvath and Peter Langfelder
}
\keyword{hplot}

