% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useNThreads.R
\name{allocateJobs}
\alias{allocateJobs}
\title{Divide tasks among workers}
\usage{
allocateJobs(nTasks, nWorkers)
}
\arguments{
\item{nTasks}{number of tasks to be divided}

\item{nWorkers}{number of workers}
}
\value{
A list with one component per worker giving the task indices to be
worked on by each worker. If there are more workers than tasks, the tasks
for the extra workers are 0-length numeric vectors.
}
\description{
This function calculates an even splitting of a given number of tasks among
a given number of workers (threads).
}
\details{
Tasks are labeled consecutively 1,2,..., \code{nTasks}. The tasks are split
in contiguous blocks as evenly as possible.
}
\examples{

allocateJobs(10, 3)
allocateJobs(2,4)

}
\author{
Peter Langfelder
}
\keyword{misc}
