% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empiricalBayesLM.R
\name{bicovWeights}
\alias{bicovWeights}
\title{Weights used in biweight midcovariance}
\usage{
bicovWeights(x, pearsonFallback = TRUE, maxPOutliers = 1,
  outlierReferenceWeight = 0.5625, defaultWeight = 0)
}
\description{
Weights used in biweight midcovariance
}
