% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchSplit.R
\name{branchEigengeneDissim}
\alias{branchEigengeneDissim}
\alias{multiSet.branchEigengeneDissim}
\title{Branch dissimilarity based on eigennodes (eigengenes).}
\usage{
branchEigengeneDissim(expr, branch1, branch2, corFnc = cor,
  corOptions = list(use = "p"), signed = TRUE, ...)
}
\arguments{
\item{expr}{Expression data.}

\item{branch1}{Branch 1.}

\item{branch2}{Branch 2.}

\item{corFnc}{Correlation function.}

\item{corOptions}{Other arguments to the correlation function.}

\item{signed}{Should the network be considered signed?}

\item{\dots}{Other arguments for compatibility; currently unused.}
}
\value{
A single number or a list containing details of the calculation.
}
\description{
Calculation of branch dissimilarity based on eigennodes (eigengenes) in
single set and multi-data situations. This function is used as a plugin for
the dynamicTreeCut package and the user should not call this function
directly. This function is experimental and subject to change.
}
\author{
Peter Langfelder
}
\keyword{misc}

