% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchSplit.R
\name{branchEigengeneDissim}
\alias{branchEigengeneDissim}
\alias{multiSet.branchEigengeneDissim}
\title{Branch dissimilarity based on eigennodes (eigengenes).}
\usage{
branchEigengeneDissim(expr, branch1, branch2, corFnc = cor,
  corOptions = list(use = "p"), signed = TRUE, ...)
}
\description{
Branch dissimilarity based on eigennodes (eigengenes).
}
