% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchSplit.R
\name{branchSplitFromStabilityLabels}
\alias{branchSplitFromStabilityLabels}
\title{Branch split (dissimilarity) statistic derived from labels determined from a
stability study}
\usage{
branchSplitFromStabilityLabels(branch1, branch2, stabilityLabels,
  ignoreLabels = 0, ...)
}
\description{
Branch split (dissimilarity) statistic derived from labels determined from a
stability study
}
