% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coClustering.R
\name{clusterCoef}
\alias{clusterCoef}
\title{Clustering coefficient calculation}
\usage{
clusterCoef(adjMat)
}
\arguments{
\item{adjMat}{adjacency matrix}
}
\value{
A vector of clustering coefficients for each node.
}
\description{
This function calculates the clustering coefficients for all nodes in the
network given by the input adjacency matrix.
}
\author{
Steve Horvath
}
\keyword{misc}
