% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{collectGarbage}
\alias{collectGarbage}
\title{Iterative garbage collection.}
\usage{
collectGarbage()
}
\value{
None.
}
\description{
Performs garbage collection until free memory idicators show no change.
}
\author{
Steve Horvath
}
\keyword{utilities}
