% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileC.R
\name{colQuantileC}
\alias{colQuantileC}
\alias{rowQuantileC}
\title{Fast colunm- and row-wise quantile of a matrix.}
\usage{
colQuantileC(data, p)
}
\arguments{
\item{data}{a numerical matrix column-wise quantiles are desired. Missing
values are removed.}

\item{p}{a single probability at which the quantile is to be calculated.}
}
\value{
A vector of length equal the number of columns (for
\code{colQuantileC}) or rows (for \code{rowQuantileC}) in \code{data}
containing the column- or row-wise quantiles.
}
\description{
Fast calculation of column- and row-wise quantiles of a matrix at a single
probability. Implemented via compiled code, it is much faster than the
equivalent \code{apply(data, 2, quantile, prob = p)}.
}
\details{
At present, only one quantile type is implemented, namely the default type 7
used by R.
}
\examples{

data <- rnorm(100)
p <- 0.1
colQuantileC(data, p)

}
\seealso{
\code{\link[stats]{quantile}}
}
\author{
Peter Langfelder
}
\keyword{misc}
