% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corFunctions.R
\name{cor}
\alias{cor}
\alias{cor1}
\alias{corFast}
\title{Correlation functions optimized}
\usage{
cor(x, y = NULL, use = "all.obs", method = c("pearson", "kendall",
  "spearman"), quick = 0, cosine = FALSE, cosineX = cosine,
  cosineY = cosine, drop = FALSE, nThreads = 0, verbose = 0,
  indent = 0)

cor1(x, use = "all.obs", verbose = 0, indent = 0)

corFast(x, y = NULL, use = "all.obs", quick = 0, nThreads = 0,
  verbose = 0, indent = 0)
}
\description{
Calculates teh correlation of the objects

Calculates teh correlation of the objects
}

