% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{corPredictionSuccess}
\alias{corPredictionSuccess}
\title{Qunatification of success of gene screening}
\usage{
corPredictionSuccess(corPrediction, corTestSet, topNumber = 100)
}
\arguments{
\item{corPrediction}{a vector or a matrix of prediction statistics}

\item{corTestSet}{correlation or other statistics on test set}

\item{topNumber}{a vector of the number of top genes to consider}
}
\value{
\item{meancorTestSetOverall }{ difference of
\code{meancorTestSetPositive} and \code{meancorTestSetNegative} below }
\item{meancorTestSetPositive}{ mean \code{corTestSet} on top genes with
positive \code{corPrediction} } \item{meancorTestSetNegative}{ mean
\code{corTestSet} on top genes with negative \code{corPrediction} } ...
}
\description{
This function calculates the success of gene screening.
}
\details{
For each column in \code{corPrediction}, the function evaluates the mean
\code{corTestSet} for the number of top genes (ranked by the column in
\code{corPrediction}) given in \code{topNumber}. The higher the mean
\code{corTestSet} (for positive \code{corPrediction}) or negative (for
negative \code{corPrediction}), the more successful the prediction.
}
\author{
Steve Horvath
}
\seealso{
\code{\link{relativeCorPredictionSuccess}}
}
\keyword{misc}

