% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{corPvalueStudent}
\alias{corPvalueStudent}
\title{Student asymptotic p-value for correlation}
\usage{
corPvalueStudent(cor, nSamples)
}
\arguments{
\item{cor}{A vector of correlation values whose corresponding p-values are
to be calculated}

\item{nSamples}{Number of samples from which the correlations were
calculated}
}
\value{
A vector of p-values of the same length as the input correlations.
}
\description{
Calculates Student asymptotic p-value for given correlations.
}
\author{
Steve Horvath and Peter Langfelder
}
\keyword{misc}

