% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorsFuncs.R
\name{displayColors}
\alias{displayColors}
\title{Show colors used to label modules}
\usage{
displayColors(colors = NULL)
}
\arguments{
\item{colors}{colors to be displayed. Defaults to all colors available for
module labeling.}
}
\value{
None.
}
\description{
The function plots a barplot using colors that label modules.
}
\details{
To see the first \code{n} colors, use argument \code{colors =
standardColors(n)}.
}
\examples{
displayColors(standardColors(10))
displayColors()
}
\author{
Peter Langfelder
}
\seealso{
\code{\link{standardColors}}
}
\keyword{misc}

