% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{goodGenes}
\alias{goodGenes}
\title{Find the good genes of a multiset object}
\usage{
goodGenes(datExpr, useSamples = NULL, useGenes = NULL, minFraction = 1/2,
  minNSamples = ..minNSamples, minNGenes = ..minNGenes, tol = NULL,
  verbose = 1, indent = 0)
}
\arguments{
\item{datExpr}{MultiSet data of expression values.}

\item{useSamples}{Select the samples to use.}

\item{useGenes}{Select the genes to use.}

\item{minFraction}{Minimal fraction to...}

\item{minNSamples}{Minimal number of samples.}

\item{minNGenes}{Minimal number of genes.}

\item{tol}{Strange option.}

\item{verbose}{Indicate the loggin output.}

\item{indent}{Indicates how much indentation you want.}
}
\description{
Find the good genes of a multiset object
}

