% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorsPalette.R
\name{greenBlackRed}
\alias{greenBlackRed}
\title{Green-black-red color sequence}
\usage{
greenBlackRed(n, gamma = 1)
}
\arguments{
\item{n}{number of colors to be returned}

\item{gamma}{color correction power}
}
\value{
A vector of colors of length \code{n}.
}
\description{
Generate a green-black-red color sequence of a given length.
}
\details{
The function returns a color vector that starts with pure green, gradually
turns into black and then to red. The power \code{gamma} can be used to
control the behaviour of the quarter- and three quarter-values (between
green and black, and black and red, respectively). Higher powers will make
the mid-colors more green and red, respectively.
}
\examples{


  par(mfrow = c(3, 1))
  \dontrun{
  displayColors(greenBlackRed(50))
  displayColors(greenBlackRed(50, 2))
  displayColors(greenBlackRed(50, 0.5))
}

}
\author{
Peter Langfelder
}
\keyword{color}
