% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{initProgInd}
\alias{initProgInd}
\alias{updateProgInd}
\title{Progress indicators}
\usage{
initProgInd(leadStr = "..", trailStr = "", quiet = !interactive())

updateProgInd(newFrac, progInd, quiet = !interactive())
}
\description{
Describes how much/how long does it takes

Should work on all OS to indicate the evolution of indicator
}

