% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiData.R
\name{keepCommonProbes}
\alias{keepCommonProbes}
\title{Keep probes that are shared among given data sets}
\usage{
keepCommonProbes(multiExpr, orderBy = 1)
}
\arguments{
\item{multiExpr}{expression data in the multi-set format (see
\code{\link{checkSets}}). A vector of lists, one per set. Each set must
contain a component \code{data} that contains the expression data, with rows
corresponding to samples and columns to genes or probes.}

\item{orderBy}{index of the set by which probes are to be ordered.}
}
\value{
Expression data in the same format as the input data, containing
only common probes.
}
\description{
This function strips out probes that are not shared by all given data sets,
and orders the remaining common probes using the same order in all sets.
}
\examples{
data1 <- matrix(rnorm(100L), 20L, 5L)
data2 <- matrix(rnorm(50L), 10L, 5L)
colnames(data1) <- LETTERS[1:5]
colnames(data2) <- LETTERS[2:6]
md <- multiSet(Set1 = data1, Set2 = data2)
keepCommonProbes(md)
}
\seealso{
\code{\link{checkSets}}
}
\author{
Peter Langfelder
}
\keyword{misc}
