% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiData.R
\name{list2multiData}
\alias{list2multiData}
\alias{multiData2list}
\title{Convert a list to a multiData structure and vice-versa.}
\usage{
list2multiData(data)
}
\arguments{
\item{data}{A list to be converted to a multiData structure.}

\item{multiData}{A multiData structure to be converted to a list.}
}
\value{
For \code{list2multiData}, a multiData structure; for
\code{multiData2list}, the corresponding list.
}
\description{
\code{list2multiData} converts a list to a multiData structure;
\code{multiData2list} does the inverse.
}
\details{
A multiData structure is a vector of lists (one list for each set) where
each list has a component \code{data} containing some useful information.
}
\author{
Peter Langfelder
}
\keyword{misc}

