% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiData.R
\name{mtd.setAttr}
\alias{mtd.setAttr}
\title{Set attributes on each component of a multiData structure}
\usage{
mtd.setAttr(multiData, attribute, valueList)
}
\arguments{
\item{multiData}{A multiData structure.}

\item{attribute}{Name for the attribute to be set}

\item{valueList}{List that gives the attribute value for each set in the
multiData structure.}
}
\value{
The input multiData with the attribute set on each \code{data}
component.
}
\description{
Set attributes on each \code{data} component of a multiData structure
}
\author{
Peter Langfelder
}
\seealso{
\code{\link{multiData}} to create a multiData structure;

\code{isMultiData} for a description of the multiData structure.
}
\keyword{misc}

