% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiData.R
\name{multiUnion}
\alias{multiIntersect}
\alias{multiUnion}
\title{Union and intersection of multiple sets}
\usage{
multiUnion(setList)

multiIntersect(setList)
}
\arguments{
\item{setList}{A list containing the sets to be performed upon.}
}
\value{
The union or intersection of the given sets.
}
\description{
Union and intersection of multiple sets. These function generalize the
standard functions \code{\link{union}} and \code{\link{intersect}}.
}
\examples{
data1 <- matrix(rnorm(100L), 20L, 5L)
data2 <- matrix(rnorm(50L), 10L, 5L)
colnames(data1) <- LETTERS[1:5]
colnames(data2) <- LETTERS[2:6]
multiUnion(list(data1, data2))
multiIntersect(list(data1, data2))
}
\author{
Peter Langfelder
}
\seealso{
The "standard" functions \code{\link{union}} and \code{\link{intersect}}.
}
\keyword{misc}

