% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{multiUnion}
\alias{multiIntersect}
\alias{multiUnion}
\title{Union and intersection of multiple sets}
\usage{
multiUnion(setList)

multiIntersect(setList)
}
\arguments{
\item{setList}{A list containing the sets to be performed upon.}
}
\value{
The union or intersection of the given sets.
}
\description{
Union and intersection of multiple sets. These function generalize the
standard functions \code{\link{union}} and \code{\link{intersect}}.
}
\author{
Peter Langfelder
}
\seealso{
The "standard" functions \code{\link{union}} and \code{\link{intersect}}.
}
\keyword{misc}

