% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusTOM.R
\name{newConsensusTree}
\alias{newConsensusTree}
\alias{ConsensusTree}
\title{Create a new consensus tree}
\usage{
newConsensusTree(consensusOptions = newConsensusOptions(), inputs,
  analysisName = NULL)
}
\arguments{
\item{consensusOptions}{An object of class \code{ConsensusOptions}, usually
obtained by calling \code{\link{newConsensusOptions}}.}

\item{inputs}{A vector (or list) of inputs. Each component can be either a
character string giving a names of a data set, or another
\code{ConsensusTree} object.}

\item{analysisName}{Optional specification of a name for this consensus
analysis. While this has no effect on the actual consensus calculation, some
functions use this character string to make certain file names unique.}
}
\value{
A list with class set to \code{"ConsensusTree"} with these
components: \code{consensusOptions}A copy of the input
\code{consensusOptions}. \code{inputs}A copy of the input \code{inputs}.
\code{analysisName}A copy of the input \code{analysisName}.
}
\description{
This function creates a new consensus tree, a class for representing
"recipes" for hierarchical consensus calculations.
}
\details{
Consensus trees specify a "recipe" for the calculation of hierarchical
consensus in \code{\link{hierarchicalConsensusCalculation}} and other
functions.
}
\seealso{
\code{\link{hierarchicalConsensusCalculation}} for hierarchical
consensus calculation for which a \code{ConsensusTree} object specifies the
recipe
}
\author{
Peter Langfelder
}
\keyword{misc}
