% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleLabels.R
\name{normalizeLabels}
\alias{normalizeLabels}
\title{Transform numerical labels into normal order.}
\usage{
normalizeLabels(labels, keepZero = TRUE)
}
\arguments{
\item{labels}{Numerical labels.}

\item{keepZero}{If \code{TRUE} (the default), labels 0 are preserved.}
}
\value{
A vector of the same length as input, containing the normalized
labels.
}
\description{
Transforms numerical labels into normal order, that is the largest group
will be labeled 1, next largest 2 etc. Label 0 is optionally preserved.
}
\examples{
labels <- c(0, 1, rep(2, 2))
labels
normalizeLabels(labels)
normalizeLabels(labels, FALSE)
}
\author{
Peter Langfelder, \email{Peter.Langfelder@gmail.com}
}
\keyword{misc}
