% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-multiData.R
\name{nSets}
\alias{nSets}
\title{Count the number of sets}
\usage{
nSets(multiData, ...)
}
\arguments{
\item{multiData}{List of Expression data, creating a multiset object.}
}
\value{
The number of sets of a multisets object
}
\description{
From a given multiset object calculates the numbers of sets present
}

