% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchLabels.R
\name{overlapTable}
\alias{overlapTable}
\title{Calculate overlap of modules}
\usage{
overlapTable(labels1, labels2, na.rm = TRUE, ignore = NULL)
}
\arguments{
\item{labels1}{a vector containing module labels.}

\item{labels2}{a vector containing module labels to be compared to
\code{labels1}.}

\item{na.rm}{logical: should entries missing in either \code{labels1} or
\code{labels2} be removed?}

\item{ignore}{an optional vector giving label levels that are to be ignored.}
}
\value{
A list with the following components: \item{countTable}{a matrix
whose rows correspond to modules (unique labels) in \code{labels1} and whose
columns correspond to modules (unique labels) in \code{labels2}, giving the
number of objects in the intersection of the two respective modules. }

\item{pTable}{a matrix whose rows correspond to modules (unique labels) in
\code{labels1} and whose columns correspond to modules (unique labels) in
\code{labels2}, giving Fisher's exact test significance p-values for the
overlap of the two respective modules. }
}
\description{
The function calculates overlap counts and Fisher exact test p-values for
the given two sets of module assignments.
}
\examples{
set.seed(1)
nModules <- 10
nGenes <- 1000
labels1 <- sample(c(1:nModules), nGenes, replace = TRUE)
labels2 <- sample(c(1:nModules), nGenes, replace = TRUE)
overlapTable(labels1, labels2)
}
\seealso{
\code{\link{fisher.test}}, \code{\link{matchLabels}}
}
\author{
Peter Langfelder
}
\keyword{misc}
