% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{prepComma}
\alias{prepComma}
\title{Prepend a comma to a non-empty string}
\usage{
prepComma(s)
}
\arguments{
\item{s}{Character string.}
}
\value{
If \code{s} is non-empty, returns \code{paste(",", s)}, otherwise
returns s.
}
\description{
Utility function that prepends a comma before the input string if the string
is non-empty.
}
\examples{

prepComma("abc")
prepComma("")

}
\author{
Peter Langfelder
}
\keyword{misc}
