% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{prependZeros}
\alias{prependZeros}
\title{Pad numbers with leading zeros to specified total width}
\usage{
prependZeros(x, width = max(nchar(x)))
}
\arguments{
\item{x}{Vector of numbers to be padded.}

\item{width}{Width to pad the numbers to.}
}
\value{
Character vector with the 0-padded numbers.
}
\description{
This function pads the specified numbers with zeros to a specified total
width.
}
\examples{

prependZeros(1:10)
prependZeros(1:10, 4)

}
\author{
Peter Langfelder
}
\keyword{misc}
