% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{relativeCorPredictionSuccess}
\alias{relativeCorPredictionSuccess}
\title{Compare prediction success}
\usage{
relativeCorPredictionSuccess(corPredictionNew, corPredictionStandard,
  corTestSet, topNumber = 100)
}
\arguments{
\item{corPredictionNew}{Matrix of predictor statistics}

\item{corPredictionStandard}{Reference presdictor statistics}

\item{corTestSet}{Correlations of predictor variables with trait in test set}

\item{topNumber}{A vector giving the numbers of top genes to consider}
}
\value{
Data frame with components \item{topNumber}{copy of the input
\code{topNumber}} \item{kruskalp}{Kruskal-Wallis p-values}
}
\description{
Compare prediction success of several gene screening methods.
}
\author{
Steve Horvath
}
\seealso{
\code{\link{corPredictionSuccess}}
}
\keyword{misc}

