% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorsFuncs.R
\name{standardColors}
\alias{standardColors}
\title{Colors this library uses for labeling modules.}
\usage{
standardColors(n = NULL)
}
\arguments{
\item{n}{Number of colors requested. If \code{NULL}, all (approx. 450)
colors will be returned. Any other invalid argument such as less than one or
more than maximum (\code{length(standardColors())}) will trigger an error.}
}
\value{
A vector of character color names of the requested length.
}
\description{
Returns the vector of color names in the order they are assigned by other
functions in this library.
}
\examples{
standardColors(10)
}
\author{
Peter Langfelder, \email{Peter.Langfelder@gmail.com}
}
\keyword{color}
\keyword{misc}

