% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{stdErr}
\alias{stdErr}
\title{Standard error of the mean of a given vector.}
\usage{
stdErr(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
Standard error of the mean of x.
}
\description{
Returns the standard error of the mean of a given vector. Missing values are
ignored.
}
\author{
Steve Horvath
}
\keyword{misc}

