% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transposeBigData.R
\name{transposeBigData}
\alias{transposeBigData}
\title{Transpose a big matrix or data frame}
\usage{
transposeBigData(x, blocksize = 20000)
}
\arguments{
\item{x}{a matrix or data frame}

\item{blocksize}{a positive integer larger than 1, which determines the
block size. Default is 20k.}
}
\description{
This transpose command partitions a big matrix (or data frame) into blocks
and applies the t() function to each block separately.
}
\details{
Assume you have a very large matrix with say 500k columns. In this case, the
standard transpose function of R \code{t()} can take a long time. Solution:
Split the original matrix into sub-matrices by dividing the columns into
blocks. Next apply \code{t()} to each sub-matrix. The same holds if the
large matrix contains a large number of rows.  The function
\code{transposeBigData} automatically checks whether the large matrix
contains more rows or more columns. If the number of columns is larger than
or equal to the number of rows then the block wise splitting will be applied
to columns otherwise to the rows.
}
\note{
This function can be considered a wrapper of \code{\link{t}()}
}
\examples{

x=data.frame(matrix(1:10000,nrow=4,ncol=2500))
dimnames(x)[[2]]=paste("Y",1:2500,sep="")
xTranspose=transposeBigData(x)
x[1:4,1:4]
xTranspose[1:4,1:4]


}
\author{
Steve Horvath, UCLA
}
\references{
Any linear algebra book will explain the transpose.
}
\seealso{
The standard function \code{\link{t}} .
}
\keyword{misc}

